Move hooks to package
