+++ packages/hooks/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@app/hooks",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "@app/api": "1.0.0",
+    "@sentry/browser": "^4.6.5",
+    "io-ts": "^1.8.3",
+    "next": "^8.0.3",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.5",
+    "react-dom": "^16.8.5",
+    "react-intl": "^2.8.0",
+    "react-relay": "^3.0.0",
+    "relay-runtime": "^3.0.0"
+  },
+  "devDependencies": {}
+}
