+++ packages/web/components/DeleteWeb.tsx
@@ -2,10 +2,10 @@ import Router from 'next/router';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { graphql } from 'react-relay';
+import useConfirm from '@app/hooks/useConfirm';
+import useMutation from '@app/hooks/useMutation';
+import { AppHref } from '@app/hooks/useAppHref';
 import { DeleteWebMutation } from '../generated/DeleteWebMutation.graphql';
-import useConfirm from '../hooks/useConfirm';
-import useMutation from '../hooks/useMutation';
-import { AppHref } from '../types';
 import Button from './Button';
 
 const mutation = graphql`
