+++ packages/web/components/ValidationError.tsx
@@ -2,7 +2,7 @@ import { EmailError, Max140CharsError, PasswordError } from '@app/api/types';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { Text } from 'react-native';
-import useAppContext from '../hooks/useAppContext';
+import useAppContext from '@app/hooks/useAppContext';
 
 // '%future added value' is reminder that GraphQL services often expand in
 // capabilities and may return new enum values. It's generated by Relay.
