+++ packages/web/pages/index.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { FormattedRelative } from 'react-intl';
 import { Text, View } from 'react-native';
 import { createFragmentContainer, graphql } from 'react-relay';
+import useAppContext from '@app/hooks/useAppContext';
+import usePageTitles from '@app/hooks/usePageTitles';
 import CreateWeb from '../components/CreateWeb';
 import Layout from '../components/Layout';
 import Link from '../components/Link';
 import { pagesQuery } from '../generated/pagesQuery.graphql';
-import useAppContext from '../hooks/useAppContext';
-import usePageTitles from '../hooks/usePageTitles';
 
 type Viewer = NonNullable<pagesQuery['viewer']>;
